requirements.txt
Flask>=2.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
Pillow>=8.0.0
python-dotenv>=0.19.0
